Added Layers, their Stacks and their event handling

+++ CCC/src/CCC/Events/MouseEvent.h
@@ -54,7 +54,7 @@ namespace CCC
 	class CCC_API MouseButtonEvent : public Event
 	{
 	public:
-		inline float GetMouseButton() const { return m_Button; }
+		inline int GetMouseButton() const { return m_Button; }
 		
 		EVENT_CLASS_CATEGORY(EventCategoryMouse | EventCategoryInput)
